tests: clean up some resource leaks for system tests
