+++ handwritten/bigtable/system-test/bigtable.ts
@@ -35,7 +35,7 @@ import {Family} from '../src/family.js';
 import {Row} from '../src/row.js';
 import {Table} from '../src/table.js';
 import {RawFilter} from '../src/filter';
-import {generateId, PREFIX} from './common';
+import {generateId, reapBackups, reapInstances} from './common';
 import {BigtableTableAdminClient} from '../src/v2';
 import {ServiceError} from 'google-gax';
 import {BigtableDate, QueryResultRow} from '../src/execute-query/values';
@@ -56,49 +56,8 @@ describe('Bigtable', () => {
   const CLUSTER_ID = generateId('cluster');
   const CLUSTER_ID_HDD = generateId('cluster');
 
-  async function reapBackups(instance: Instance) {
-    try {
-      const [backups] = await instance.getBackups();
-      for (const backup of backups) {
-        try {
-          await backup.delete({timeout: 50 * 1000});
-        } catch (e) {
-          console.log(`Error deleting backup: ${backup.id}: ${e}`);
-        }
-      }
-    } catch (e) {
-      console.error(`Error listing backups from ${instance.name}: ${e}`);
-    }
-  }
-
-  async function reapInstances() {
-    const [instances] = await bigtable.getInstances();
-    const testInstances = instances
-      .filter(i => i.id.match(PREFIX))
-      .filter(i => {
-        const timeCreated = i.metadata!.labels!.time_created as {} as Date;
-        // Only delete stale resources.
-        const oneHourAgo = new Date(Date.now() - 3600000);
-        return !timeCreated || timeCreated <= oneHourAgo;
-      });
-    // need to delete backups first due to instance deletion precondition
-    const deleteBackupPromises = testInstances.map(instance =>
-      reapBackups(instance),
-    );
-    for (const backupPromise of deleteBackupPromises) {
-      await backupPromise;
-    }
-    for (const instance of testInstances) {
-      try {
-        await instance.delete();
-      } catch (e) {
-        console.log(`Error deleting instance: ${instance.id}`);
-      }
-    }
-  }
-
   before(async () => {
-    await reapInstances();
+    await reapInstances(bigtable);
     const [, operation] = await INSTANCE.create({
       clusters: [
         {
@@ -140,22 +99,20 @@ describe('Bigtable', () => {
   });
 
   after(async () => {
-    const q = [];
     const instances = [INSTANCE, DIFF_INSTANCE, CMEK_INSTANCE, INSTANCE_HDD];
 
     // need to delete backups first due to instance deletion precondition
     await Promise.all(instances.map(instance => reapBackups(instance)));
     await Promise.all(
-      instances.map(instance => {
-        q.push(async () => {
-          try {
-            await instance.delete();
-          } catch (e) {
-            console.log(`Error deleting instance: ${instance.id}`);
-          }
-        });
+      instances.map(async instance => {
+        try {
+          await instance.delete();
+        } catch (e) {
+          console.log(`Error deleting instance: ${instance.id}`);
+        }
       }),
     );
+    await reapInstances(bigtable);
   });
 
   describe('instances', () => {
