+++ handwritten/bigtable/system-test/client-side-metrics-all-methods.ts
@@ -38,7 +38,7 @@ import {ClientOptions} from 'google-gax';
 import {ClientSideMetricsConfigManager} from '../src/client-side-metrics/metrics-config-manager';
 import {MetricServiceClient} from '@google-cloud/monitoring';
 import {MethodName} from '../src/client-side-metrics/client-side-metrics-attributes';
-import {generateId} from './common';
+import {generateId, reapInstances} from './common';
 
 const SECOND_PROJECT_ID = 'cfdb-sdk-node-tests';
 const instanceId1 = generateId('instance');
@@ -357,6 +357,7 @@ describe('Bigtable/ClientSideMetricsAllMethods', () => {
   let defaultProjectId: string;
 
   before(async () => {
+    await reapInstances(new Bigtable());
     /*
     For both the default project and the secondary project we need to create
     instances with some data in them so that the tests can collect all the
@@ -413,6 +414,7 @@ describe('Bigtable/ClientSideMetricsAllMethods', () => {
         console.warn('The instance has been deleted already');
       }
     }
+    await reapInstances(new Bigtable());
   });
 
   describe('Bigtable/ClientSideMetricsToGCM', () => {
